Fix type
git-svn-id: svn://svn.code.sf.net/p/tass64/code/trunk@2430 111d7451-93fa-4e81-90d4-8039f76616e9

+++ typeobj.c
@@ -65,7 +65,7 @@ static FAST_CALL bool same(const Obj *o1, const Obj *o2) {
 
 static MUST_CHECK Error *hash(Obj *o1, int *hs, linepos_t UNUSED(epoint)) {
     Type *v1 = (Type *)o1;
-    *hs = v1->type;
+    *hs = (int)v1->type;
     return NULL;
 }
